Fikser tester og bruker ny kontrakt.
