+++ pom.xml
@@ -23,7 +23,7 @@
         <springfox.version>2.9.2</springfox.version>
         <common-java-modules.version>1.2019.11.06-15.14-6ca860e6cead</common-java-modules.version>
         <felles.version>1.20200427140802_3cc4a07</felles.version>
-        <kontrakt.version>2.0-SNAPSHOT</kontrakt.version>
+        <kontrakt.version>2.0_20200428192228_09123cf</kontrakt.version>
         <fasterxml.version>2.10.1</fasterxml.version>
         <cxf.version>3.3.4</cxf.version>
         <token-validation-spring.version>1.1.2</token-validation-spring.version>
