+++ src/main/java/no/nav/familie/integrasjoner/client/rest/DokarkivLogiskVedleggRestClient.kt
@@ -15,7 +15,7 @@ import java.net.URI
 
 @Component
 class DokarkivLogiskVedleggRestClient(@Value("\${DOKARKIV_V1_URL}") private val dokarkivUrl: URI,
-                                      @Qualifier("jwtBearer") private val restOperations: RestOperations)
+                                      @Qualifier("sts") private val restOperations: RestOperations)
     : AbstractRestClient(restOperations, "dokarkiv.logiskvedlegg.opprett") {
 
     private val slettVedleggClient = SlettLogiskVedleggClient(restOperations)
